+++ Plugins/DotNET/NWN/Internal/Bootstrap.cs
@@ -129,6 +129,10 @@ namespace NWN
         public delegate IntPtr nwnxPopItemPropertyDelegate();
         [SuppressUnmanagedCodeSecurity][UnmanagedFunctionPointer(CallingConvention.Cdecl)]
         public delegate void nwnxCallFunctionDelegate();
+        [SuppressUnmanagedCodeSecurity][UnmanagedFunctionPointer(CallingConvention.Cdecl)]
+        public delegate IntPtr RequestHookDelegate(IntPtr address, IntPtr managedFuncPtr, int priority);
+        [SuppressUnmanagedCodeSecurity][UnmanagedFunctionPointer(CallingConvention.Cdecl)]
+        public delegate void ReturnHookDelegate(IntPtr hook);
 
         [StructLayout(LayoutKind.Sequential)]
         public readonly struct BootstrapArgs
@@ -171,6 +175,8 @@ namespace NWN
             public readonly nwnxPopItemPropertyDelegate           nwnxPopItemProperty;
             public readonly nwnxCallFunctionDelegate              nwnxCallFunction;
             public readonly GetNWNXExportedGlobalsDelegate        GetNWNXExportedGlobals;
+            public readonly RequestHookDelegate                   RequestHook;
+            public readonly ReturnHookDelegate                    ReturnHook;
         }
         public static BootstrapArgs NativeFunctions;
         private static AllHandlers _handlers;